+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [#121] Implementation of an LRU cache for Helm charts to prevent peak loads in the Harbor registry
+- [#121] Use exponential backoff with cap for reconciliation of components 
+- [#121] Update go version to 1.26.8 and dependencies
 
 ## [v1.14.1] - 2026-07-23
 ### Added